make gammapy.modeling.models.spectrum work
