+++ gammapy/astro/darkmatter/profiles.py
@@ -4,7 +4,7 @@ import abc
 import numpy as np
 import astropy.units as u
 from ...utils.fitting import Parameter, Parameters
-from ...spectrum.utils import integrate_spectrum
+from gammapy.modeling.models.spectrum.utils import integrate_spectrum
 
 __all__ = [
     "DMProfile",
