+++ gammapy/modeling/models/spectrum/cosmic_ray.py
