+++ gammapy/modeling/models/spectrum/crab.py
