+++ gammapy/modeling/models/spectrum/tests/test_crab.py
@@ -1,8 +1,8 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 import pytest
 import astropy.units as u
-from ...utils.testing import assert_quantity_allclose
-from ...spectrum import create_crab_spectral_model
+from gammapy.utils.testing import assert_quantity_allclose
+from ..crab import create_crab_spectral_model
 
 CRAB_SPECTRA = [
     dict(
