+++ gammapy/modeling/models/spectrum/utils.py
@@ -2,106 +2,7 @@
 import numpy as np
 from astropy.units import Quantity
 
-__all__ = ["SpectrumEvaluator", "integrate_spectrum"]
-
-
-class SpectrumEvaluator:
-    """Calculate number of predicted counts (``npred``).
-
-    The true and reconstructed energy binning are inferred from the provided IRFs.
-
-    Parameters
-    ----------
-    model : `~gammapy.spectrum.image.SpectralModel`
-        Spectral model
-    aeff : `~gammapy.irf.EffectiveAreaTable`
-        EffectiveArea
-    edisp : `~gammapy.irf.EnergyDispersion`, optional
-        EnergyDispersion
-    livetime : `~astropy.units.Quantity`
-        Observation duration (may be contained in aeff)
-    e_true : `~astropy.units.Quantity`, optional
-        Desired energy axis of the prediced counts vector if no IRFs are given
-
-    Examples
-    --------
-    Calculate predicted counts in a desired reconstruced energy binning
-
-    .. plot::
-        :include-source:
-
-        from gammapy.irf import EnergyDispersion, EffectiveAreaTable
-        from gammapy.spectrum import image, SpectrumEvaluator
-        import numpy as np
-        import astropy.units as u
-        import matplotlib.pyplot as plt
-
-        e_true = np.logspace(-2, 2.5, 109) * u.TeV
-        e_reco = np.logspace(-2, 2, 73) * u.TeV
-
-        aeff = EffectiveAreaTable.from_parametrization(energy=e_true)
-        edisp = EnergyDispersion.from_gauss(e_true=e_true, e_reco=e_reco,
-                                            sigma=0.3, bias=0)
-
-        model = image.PowerLaw(index=2.3,
-                                amplitude="2.5e-12 cm-2 s-1 TeV-1",
-                                reference="1 TeV")
-
-        livetime = 1 * u.h
-
-        predictor = SpectrumEvaluator(model=model,
-                                      aeff=aeff,
-                                      edisp=edisp,
-                                      livetime=livetime)
-        predictor.compute_npred().plot_hist()
-        plt.show()
-    """
-
-    def __init__(self, model, aeff=None, edisp=None, livetime=None, e_true=None):
-        self.model = model
-        self.aeff = aeff
-        self.edisp = edisp
-        self.livetime = livetime
-
-        if aeff is not None:
-            e_true = self.aeff.energy.edges
-
-        self.e_true = e_true
-        self.e_reco = None
-
-    def compute_npred(self):
-        integral_flux = self.model.integral(
-            emin=self.e_true[:-1], emax=self.e_true[1:], intervals=True
-        )
-
-        true_counts = self.apply_aeff(integral_flux)
-        return self.apply_edisp(true_counts)
-
-    def apply_aeff(self, integral_flux):
-        if self.aeff is not None:
-            cts = integral_flux * self.aeff.data.data
-        else:
-            cts = integral_flux
-
-        # Multiply with livetime if not already contained in aeff or model
-        if cts.unit.is_equivalent("s-1"):
-            cts *= self.livetime
-
-        return cts.to("")
-
-    def apply_edisp(self, true_counts):
-        from . import CountsSpectrum
-
-        if self.edisp is not None:
-            cts = self.edisp.apply(true_counts)
-            self.e_reco = self.edisp.e_reco.edges
-        else:
-            cts = true_counts
-            self.e_reco = self.e_true
-
-        return CountsSpectrum(
-            data=cts, energy_lo=self.e_reco[:-1], energy_hi=self.e_reco[1:]
-        )
+__all__ = ["integrate_spectrum"]
 
 
 def integrate_spectrum(func, xmin, xmax, ndecade=100, intervals=False):
