+++ gammapy/spectrum/__init__.py
@@ -1,9 +1,6 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 """1D spectrum analysis."""
-from .utils import *
 from .core import *
-from .cosmic_ray import *
-from .crab import *
 from .dataset import *
 from .flux_point import *
 from .extract import *
