+++ gammapy/time/lightcurve.py
@@ -4,7 +4,7 @@ import numpy as np
 import astropy.units as u
 from astropy.table import Table
 from astropy.time import Time
-from ..spectrum.utils import SpectrumEvaluator
+from gammapy.spectrum.core import SpectrumEvaluator
 from ..stats.poisson import excess_error, excess_ul_helene
 from ..utils.scripts import make_path
 from ..utils.table import table_from_row_data
